ADDED: Spell icon outline and size change during cast
Added a casting bool and a getter for it in the SpellCastingController.cs as well as functionality in the Update() for the PlayerHud.cs for the spell icon to grow in size while spell is being cast and for it to decrease in size during cooldown. Maximum size increase and increase and decrease speed are settable in the editor.