FIXED: Drop rate
SelectDropRandomly() worked fine for 2 items but no more than 2. Code was changed so that it is better scalable. We get a Random value from 0 t 1 and check according to the added probability of each item. So long as total probability in drop rate table is no bigger than 1, it should work.