ADDED: Special attack that can be picked up
Lizards have 10% chance of dropping an item that adds a special attack to the player's right mouse button. When picked up, it sets an icon for the UI and acts independenlty of the simple attack. Other than that, it has the same functionalities as the simple attack, for now. A child of RuntimeDropInstance.cs was created to deal with picking up of the spell.